Moved validated forced decision to decision service and fixed tests
