+++ core-api/src/main/java/com/optimizely/ab/Optimizely.java
@@ -1185,7 +1185,7 @@ public class Optimizely implements AutoCloseable {
 
         // Check Forced Decision
         OptimizelyDecisionContext optimizelyDecisionContext = new OptimizelyDecisionContext(flag.getKey(), null);
-        DecisionResponse<Variation> forcedDecisionVariation = user.findValidatedForcedDecision(optimizelyDecisionContext);
+        DecisionResponse<Variation> forcedDecisionVariation = decisionService.validatedForcedDecision(optimizelyDecisionContext, projectConfig, user);
         decisionReasons.merge(forcedDecisionVariation.getReasons());
         if (forcedDecisionVariation.getResult() != null) {
             flagDecision = new FeatureDecision(null, forcedDecisionVariation.getResult(), FeatureDecision.DecisionSource.FEATURE_TEST);
@@ -1342,26 +1342,6 @@ public class Optimizely implements AutoCloseable {
         }
 
         return new DecisionResponse(valuesMap, reasons);
-    }
-
-    /**
-     *  Gets a variation based on flagKey and variationKey
-     *
-     * @param flagKey The flag key for the variation
-     * @param variationKey The variation key for the variation
-     * @return Returns a variation based on flagKey and variationKey, otherwise null
-     */
-    public Variation getFlagVariationByKey(String flagKey, String variationKey) {
-        Map<String, List<Variation>> flagVariationsMap = getProjectConfig().getFlagVariationsMap();
-        if (flagVariationsMap.containsKey(flagKey)) {
-            List<Variation> variations = flagVariationsMap.get(flagKey);
-            for (Variation variation : variations) {
-                if (variation.getKey().equals(variationKey)) {
-                    return variation;
-                }
-            }
-        }
-        return null;
     }
 
     /**
