+++ core-api/src/main/java/com/optimizely/ab/config/DatafileProjectConfig.java
@@ -503,6 +503,27 @@ public class DatafileProjectConfig implements ProjectConfig {
         return flagVariationsMap;
     }
 
+    /**
+     *  Gets a variation based on flagKey and variationKey
+     *
+     * @param flagKey The flag key for the variation
+     * @param variationKey The variation key for the variation
+     * @return Returns a variation based on flagKey and variationKey, otherwise null
+     */
+    @Override
+    public Variation getFlagVariationByKey(String flagKey, String variationKey) {
+        Map<String, List<Variation>> flagVariationsMap = getFlagVariationsMap();
+        if (flagVariationsMap.containsKey(flagKey)) {
+            List<Variation> variations = flagVariationsMap.get(flagKey);
+            for (Variation variation : variations) {
+                if (variation.getKey().equals(variationKey)) {
+                    return variation;
+                }
+            }
+        }
+        return null;
+    }
+
     @Override
     public String toString() {
         return "ProjectConfig{" +
